Rename ProjectBucketLink component (add External prefix)
